Move mailing tests in test_participant_sevice
